+++ client/client-netty/src/main/java/org/jboss/aerogear/sync/client/netty/WebSocketClientHandler.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.sync;
+package org.jboss.aerogear.sync.client.netty;
 
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
