+++ client/client-netty/src/main/java/org/jboss/aerogear/sync/client/netty/package-info.java
@@ -1,0 +1,23 @@
+/**
+ * JBoss, Home of Professional Open Source
+ * Copyright Red Hat, Inc., and individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * 	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * This package contains Netty Client for AeroGear Sync.
+ *
+ * @see org.jboss.aerogear.sync.client.netty.DiffSyncClient
+ */
+package org.jboss.aerogear.sync.client.netty;
