+++ server/server-netty/src/main/java/org/jboss/aerogear/sync/server/netty/StandaloneConfig.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.sync;
+package org.jboss.aerogear.sync.server.netty;
 
 public class StandaloneConfig {
