+++ server/server-xmpp/src/main/java/org/jboss/aerogear/sync/server/gcm/GcmMessages.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.sync;
+package org.jboss.aerogear.sync.server.gcm;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import org.jboss.aerogear.sync.diffmatchpatch.JsonMapper;
